Fix app crash after screen rotation
data class ProductCategory is not Parcelable so the state cannot
be saved. When the screen is rotated, android will rebuild activity,
this causes the state in compose reinitialize. So, if the object is
not parcelable, the data will not be saved. To safely save the data,
make the ProductCategory class be parcelable.