Add ProductDetails data class
This class is used to easily show the value on UI.
Product stores category id and a list of tag id.
This makes it is difficult to get the name of category
and tags on ui layer. Thus, ProductDetails stores
ProductCategory and ProductTag so that ui layer would
be easier to get the name of them.

This patch also add AsyncImagePainter.State to handle
the state while loading a image and navigate to "productList"
instead of popBackStack()